pyproject.toml
[tool.poetry]
name = "litegrave"
version = "0.1.0"
description = ""
authors = ["Alireza <alireza.kashanipour@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.27.2"
watchdog = "^3.0.0"
loguru = "^0.7.2"
pre-commit = "^3.5.0"
ipykernel = "^6.26.0"
langchain = "^0.0.339"
langchain-experimental = "^0.0.42"
python-dotenv = "^1.0.0"
openai = "^1.3.3"
tabulate = "^0.9.0"
omegaconf = "^2.3.0"
sentence-transformers = "^2.2.2"
bs4 = "^0.0.1"
lxml = "^4.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
